+++ BiCoR/UI/WebViewController.m
@@ -39,7 +39,7 @@ int const WEB_VIEW_CONTROLLER_EDIT_USER = 2;
             dataUrl = [connection.url stringByAppendingString:[connection.userPartOfUrl stringByAppendingString:SERVER_CONNECTION_ADD_PEOPLE_PAGE_WEB]];
             break;
         case WEB_VIEW_CONTROLLER_EDIT_USER:
-            dataUrl = [connection.url stringByAppendingString:[connection.userPartOfUrl stringByAppendingString:[SERVER_CONNECTION_PERSON_PAGE_WEB stringByAppendingString:[NSString stringWithFormat:@"%i/edit", _userId]]]];
+            dataUrl = [connection.url stringByAppendingString:[connection.userPartOfUrl stringByAppendingString:[SERVER_CONNECTION_PERSON_PAGE_WEB stringByAppendingString:[NSString stringWithFormat:@"%li/edit", (long)_userId]]]];
             break;
         default:
             break;
